add: Separate scrappers
